リファクタリング：データベースおよびカラム関連のコンテキストサービスを実装 カラムフォームのエラーページを処理する ColumnPageContextService を追加
カラムデータの取得と準備を行う DbColumnContextService を導入

データベース関連ページの構成を担う DatabasePageContextService を作成

データベース情報の取得を行う DbInfoContextService を実装

複数カラムのコンテキストを扱う ColumnMultiContextService を追加

複数テーブルを扱う TableMultiContextService を導入

テーブル関連データを処理する DbTableContextService を作成

テーブル詳細ページの構築を担う TablePageContextService を追加

DataTypeResolver および関連クラスを TableColumnEntity を用いるようリファクタリング

データベースエンティティ操作用の DatabaseRepositoryService を追加

非推奨となった DatabaseService および DbTableColumnService を削除

カラム操作用の DbColumnRepositoryService を新規作成

テーブル操作用の DbTableRepositoryService を新規作成

DBInfoMapper を DTO ではなくエンティティベースで動作するよう変更

+++ src/main/java/com/example/projectmanagement/modules/databases/datastructure/dto/ColumnDtoRecord.java
@@ -2,8 +2,8 @@ package com.example.projectmanagement.modules.databases.datastructure.dto;
 
 public record ColumnDtoRecord(
 
-		Integer id,
-		Integer tableInfoId,
+		Long id,
+		Long tableInfoId,
 		String columnName,
 		String alias,
 		String dataType,
